FIX: ミドルウェアを別ファイルに切り離し
